Adding predictive search component
Created basic predictive search component using ngx-bootstrap typeahead.
Most of the features are implemented, but it's not searching the full
data, just visible elements in the currently rendered layer (to just get
the prototype working for now). Having some trouble testing the
typeahead because keyup events don't seem to be firing.