Fix Mapbox rendering on prod
There are some problems with Mapbox GL and the Angular CLI documented in
the issue, and the angular-mapbox component was contributing to that as
well. This removes that package, loads the typings for mapboxgl and
geojson, and replaces the package functionality with the fixed types.
Works with ng build --prod and then running a localhost server, and
tests still pass.