Update time slider support, layer data
Currently a work in progress, but updated the app component to only use
dataYear rather than also using censusYear now that we're just using
2010 geography. Also updated the style.json file to reflect this, and
updated it to use default values rather than filters for layers because
it's easier to manage. Attempting to change bubble layers with dataYear
and the slider, but not sure if I'm missing something or running into a
bug.