Add getUnionFeature to consolidate hover features
Instead of modifying the source data and creating larger tilesets to
account for features used in the hover layer being split in vector
tiling, this uses the spatial union function from turf.js to combine the
geographies of all features matching the name and parent-location of a
feature (which should be unique). Performance doesn't appear to suffer.