Fix zoom layer change, hover interactions
The previous PR hadn't actually fixed the layers auto-switching. Because
the UI select component fired an event on any change to the selected
element, and the onMapZoom event changed the selected level, even if
autoSwitchLayers was reset in onMapZoom the select event would un-toggle
it. Added a check in setGroupVisibility to see if the data level had
already been changed in onMapZoom, and only toggle autoSwitchLayers if
it had. Also emptying the hover layer on layer change so that state
outlines don't appear if someone zooms in while hovering and the layer
changes.