Styling for missing data in layers
Supports styling of gray bubbles and gray cross-hatch striped shapes
when data is missing (-1). Also adds the "None" option to the bubble
selector. Adds _null layers for each shape layer because fill-pattern
doesn't seem to support boolean expressions, and sprite is generated for
the fill which is filtered by the current data property being less than
0.