Fix dots with stroke for none bubbles
On the null bubble layer, even though the property it's trying to access
is fake and almost nothing displays, the circle-stroke-color property
not being related to data values meant a dot with just the stroke still
displayed. Added a property function for circle-stroke-color for all
bubble layers that really just checks for the existence of the property.