Update highlight outline, bbox behavior
Add a white outline to the highlight to make it more visible, especially
for blue lines. Also, default to bounding box if data not available
(i.e. not visible or in current layer), and then update the boundaries
as it becomes available or more refined. Fixes the subscriber to the
mapLoading event so that it's actually distinctUntilChanged, and updates
map highlight outlines whenever that happens