Connect bar year and general map year
Removes barYear on the data panel component, and updates the general
data service year instead. Also sets the ui-slider component to update
its position when the year input is changed, and doesn't call
updateHighlightFeatures when the year is updated because it can run into
errors when the tile layer switches.